+++ NEWS.md
@@ -9,6 +9,11 @@
 
 <section class="center-content">
 
+Dev
+---
+
+- Filter mappings by system and type
+
 May 31, 2025
 ---